Pass keyword args to the concrete verification objects as keywords
Previously, when `AddressFinder.verification` was invoked in practice, a hash
would be passed.  This is slightly difficult to spot in the current era, as the
warning is _silenced_ from Ruby 2.7.2, and the existing test suite did not
actually run this code.  For now, add a test that would fail under Ruby 3.0+.