Explicitly depend on `ostruct`
This is being partitioned out of the bundled gem set, and so calls such as
`require 'ostruct'` now raise warnings if `ostruct` is not an explicit
dependency of the project.  This doesn't show up in general tests of the
`addressfinder` gem, because it is brought in as an indirect development
dependency via `guard-rspec` -> `guard`.